Refactor GPTQ quantization schemes
