+++ python/sglang/srt/layers/quantization/gptq/__init__.py
@@ -1,0 +1,42 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from sglang.srt.hardware_backend.gpu.quantization.gptq_kernels import (
+    gptq_marlin_moe_repack,
+)
+
+from .gptq import (
+    GPTQAscendConfig,
+    GPTQConfig,
+    GPTQLinearAscendMethod,
+    GPTQLinearMethod,
+    GPTQMarlinConfig,
+    GPTQMarlinLinearMethod,
+    GPTQMarlinMoEMethod,
+    GPTQMoEAscendMethod,
+    check_marlin_format,
+)
+from .schemes import (
+    GPTQAscendLinearScheme,
+    GPTQLinearScheme,
+    GPTQMarlinLinearScheme,
+    GPTQMarlinMoEScheme,
+    GPTQMoEAscendScheme,
+)
+
+__all__ = [
+    "GPTQConfig",
+    "GPTQAscendConfig",
+    "GPTQMarlinConfig",
+    "GPTQLinearMethod",
+    "GPTQMoEAscendMethod",
+    "GPTQMarlinLinearMethod",
+    "GPTQLinearAscendMethod",
+    "GPTQMarlinMoEMethod",
+    "GPTQLinearScheme",
+    "GPTQAscendLinearScheme",
+    "GPTQMarlinLinearScheme",
+    "GPTQMoEAscendScheme",
+    "GPTQMarlinMoEScheme",
+    "check_marlin_format",
+    "gptq_marlin_moe_repack",
+]
