+++ python/sglang/srt/layers/quantization/gptq/schemes/__init__.py
@@ -1,0 +1,16 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from .gptq_linear import GPTQAscendLinearScheme, GPTQLinearScheme
+from .gptq_marlin import GPTQMarlinLinearScheme
+from .gptq_moe import GPTQMarlinMoEScheme, GPTQMoEAscendScheme
+from .gptq_scheme import GPTQLinearSchemeBase, GPTQMoESchemeBase
+
+__all__ = [
+    "GPTQLinearSchemeBase",
+    "GPTQMoESchemeBase",
+    "GPTQLinearScheme",
+    "GPTQAscendLinearScheme",
+    "GPTQMarlinLinearScheme",
+    "GPTQMoEAscendScheme",
+    "GPTQMarlinMoEScheme",
+]
